Added loader. Fix small bugs

+++ app/src/main/java/com/privategallery/akscorp/privategalleryandroid/Widgets/UnlockImageButton.kt
@@ -2,8 +2,11 @@ package com.privategallery.akscorp.privategalleryandroid.Widgets
 
 import android.content.Context
 import android.content.ContextWrapper
+import android.content.Intent
+import android.content.IntentFilter
 import android.os.Bundle
 import android.os.Environment
+import android.os.Handler
 import android.support.v4.view.GravityCompat
 import android.util.AttributeSet
 import android.view.View
@@ -12,6 +15,7 @@ import com.privategallery.akscorp.privategalleryandroid.Activities.MainActivity
 import com.privategallery.akscorp.privategalleryandroid.Adapters.LocalStorageGridAdapter
 import com.privategallery.akscorp.privategalleryandroid.Adapters.UnlockPreviewGridAdapter
 import com.privategallery.akscorp.privategalleryandroid.Database.LocalDatabaseAPI
+import com.privategallery.akscorp.privategalleryandroid.Dialogs.*
 import com.privategallery.akscorp.privategalleryandroid.Essentials.Image
 import com.privategallery.akscorp.privategalleryandroid.Fragments.PreviewListFragment
 import com.privategallery.akscorp.privategalleryandroid.Fragments.UnlockListFragment
@@ -20,6 +24,8 @@ import com.privategallery.akscorp.privategalleryandroid.Utilities.Utilities
 import kotlinx.android.synthetic.main.activity_main.*
 import kotlinx.android.synthetic.main.local_storage_grid_fragment.*
 import kotlinx.android.synthetic.main.preview_images_grid_fragment.*
+import kotlinx.coroutines.experimental.launch
+import java.io.FileNotFoundException
 import java.io.Serializable
 
 /**
@@ -37,24 +43,79 @@ class UnlockImageButton : ImageButton, View.OnClickListener {
 
     private fun getBaseContext() = ((context as ContextWrapper).baseContext as MainActivity)
 
+
+    private fun progressBroadcastReceiverInit(progressDialog: LoadDialog)
+    {
+        val intentFilter = IntentFilter(PROGRESS_BROADCAST_RECEIVER_TAG)
+        val mReceiver =
+            ProgressBroadcastReceiver(
+                progressDialog
+            )
+        context.registerReceiver(mReceiver, intentFilter)
+    }
+
+    private fun sentProgressToReceiver(progress: Int)
+    {
+        val intent = Intent()
+        intent.action =
+                PROGRESS_BROADCAST_RECEIVER_TAG
+        intent.putExtra(
+            CURRENT_PROGRESS_BROADCAST_RECEIVER,progress)
+        getBaseContext().sendBroadcast(intent)
+    }
+
     override fun onClick(v: View?) {
         getBaseContext().toolbar.setState(COMMON)
+
         val unlockPreviewGridAdapter = getBaseContext().main_preview_rv_grid.adapter as
                 UnlockPreviewGridAdapter
 
-        for (image in unlockPreviewGridAdapter.used) {
+        if(unlockPreviewGridAdapter.used.size == 0)
+            return
 
-            Utilities.moveFile(
-                getImagePath(image),
-                Environment.getExternalStorageDirectory().absolutePath +
-                        "/" + Environment.DIRECTORY_DCIM + "/PrivateGalleryFiles",
-                getImageName(image)
-            )
-            db.removeImageFromDatabase(image)
-            getBaseContext().currentAlbum.images.remove(image)
+        val dialog = LoadDialog()
+        dialog.showNow(getBaseContext().supportFragmentManager,
+            LOAD_DIALOG_TAG)
+
+        progressBroadcastReceiverInit(dialog)
+
+        launch {
+
+
+            var counter = 0
+            val filesCount = unlockPreviewGridAdapter.used.size.toDouble()
+
+            for (image in unlockPreviewGridAdapter.used) {
+                counter++
+
+                try {
+                    Utilities.moveFile(
+                        getImagePath(image),
+                        Environment.getExternalStorageDirectory().absolutePath +
+                                "/" + Environment.DIRECTORY_DCIM + "/PrivateGalleryFiles",
+                        getImageName(image))
+                    db.removeImageFromDatabase(image)
+                    getBaseContext().currentAlbum.images.remove(image)
+                }
+                catch (e: FileNotFoundException)
+                {
+                    db.removeImageFromDatabase(image)
+                    getBaseContext().currentAlbum.images.remove(image)
+                }
+
+                sentProgressToReceiver((counter / filesCount * 100.0).toInt())
+
+            }
+
+            getBaseContext().runOnUiThread{
+                Handler().postDelayed({
+                    dialog.dismiss()
+                    getBaseContext().fab.visibility = View.VISIBLE
+                    getBaseContext().showAlbumContent(getBaseContext().currentAlbum)
+                }, 1500)
+            }
+
         }
-        getBaseContext().fab.visibility = View.VISIBLE
-        getBaseContext().showAlbumContent(getBaseContext().currentAlbum)
     }
 
     private fun getImagePath(image: Image) =
